Yarns Loom v2.7.1

+++ yarns/arpeggiator.h
@@ -44,8 +44,8 @@ class HeldKeys;
 class SequencerArpeggiatorResult;
 
 struct Arpeggiator {
-  int8_t key_index;
-  int8_t octave;
+  int32_t key_index; // Combines with octave to select a note from the arp chord
+  int8_t octave; // Current octave of the ARP RANGE
   int8_t key_increment;
 
   void Reset();
